stockwave-backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==4.3.0
boto3==1.42.91
botocore==1.42.91
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
fastapi==0.136.0
h11==0.16.0
idna==3.11
jmespath==1.1.0
passlib==1.7.4
pydantic==2.13.0
pydantic-settings==2.13.1
pydantic_core==2.46.0
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.33.1
s3transfer==0.16.0
six==1.17.0
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
